+++ Assets/Controllers/InputController.cs
@@ -12,15 +12,7 @@ public class InputController : MonoBehaviour {
 
     void Update() {
         ShipComponent activeShipComponent = GetComponent<ShipController>().GetShipComponent(verse.ActiveShip());
-
-        Vector2 pos = Camera.main.ScreenToWorldPoint(Input.mousePosition);
-        pos.x = Mathf.Floor(pos.x + 0.5f);
-        pos.y = Mathf.Floor(pos.y + 0.5f);
-        verse.buildIndicatorEntity.Position = pos;
-
-        if( Input.GetKeyDown(KeyCode.R)) {
-            verse.buildIndicatorEntity.Rotate();
-        }
+        
 
         Vector3 localPos = Camera.main.ScreenToWorldPoint(Input.mousePosition) - activeShipComponent.transform.position;
         localPos.z = 0;
@@ -28,6 +20,13 @@ public class InputController : MonoBehaviour {
         Coordinate coord = new Coordinate(
              rotatedPos
             );
+
+        verse.buildIndicatorEntity.Position = coord.ToVector();
+
+        if( Input.GetKeyDown(KeyCode.R)) {
+            verse.buildIndicatorEntity.Rotate();
+        }
+
         if (verse.verseMode == Verse.VerseMode.Build) {
             if (Input.GetButtonDown("Fire1")) {
                 Part part = verse.registry.partRegistry.Get(verse.modeArgs[0] as string);