+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 - Initial release
 - Basic ship and parts
 - Basic Xml/LUA interface
+- Multilayer sprites
+- Build indicator fixes
+- Engine/Flameout/Robustness of Pars Parameters increased